pyproject.toml
[project]
name = "liquidsystem"
version = "0.2.0"
description = "LiquidLab Pi-hole audit dashboard — FastAPI web app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "apscheduler>=3.10,<4",
    "fastapi[standard]>=0.111.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
